Initialize GSR with actual entry block signature instead of 0.
This is a quick fix to deal with more than one function per module or pass run,
I'm not yet sure which. After instrumenting the first function, initializing
the GSR with 0 caused signature mismatch in all following functions.

Not pretty and "-instcombine" still eats most of the checking code due to
constant folding, unfortunately.